requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.11.2
Pillow==9.4.0
requests==2.28.2
